change module name to github.com/lesta-battleship/server-core
