+++ go.mod
@@ -1,4 +1,4 @@
-module lesta-battleship/server-core
+module github.com/lesta-battleship/server-core
 
 go 1.24.3
