+++ internal/game/model.go
