+++ internal/game/openCell.go
