+++ internal/game/removeShip.go
