+++ internal/game/shoot.go
