+++ internal/ws/handlers/remove_ship.go
@@ -2,9 +2,9 @@ package handlers
 
 import (
 	"errors"
-	"lesta-battleship/server-core/internal/game-core/game"
-	"lesta-battleship/server-core/internal/game-core/match"
-	"lesta-battleship/server-core/internal/game-core/transaction"
+	"github.com/lesta-battleship/server-core/internal/game"
+	"github.com/lesta-battleship/server-core/internal/match"
+	"github.com/lesta-battleship/server-core/internal/transaction"
 
 	"github.com/gorilla/websocket"
 )
